+++ libraries/AD524X/AD524X.h
@@ -2,7 +2,7 @@
 //
 //    FILE: AD524X.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.5.0
+// VERSION: 0.5.1
 // PURPOSE: I2C digital PotentioMeter AD5241 AD5242
 //    DATE: 2013-10-12
 //     URL: https://github.com/RobTillaart/AD524X
@@ -12,7 +12,7 @@
 #include "Wire.h"
 
 
-#define AD524X_LIB_VERSION    (F("0.5.0"))
+#define AD524X_LIB_VERSION    (F("0.5.1"))
 
 
 #define AD524X_OK             0
@@ -29,6 +29,7 @@ public:
 
   bool    begin();
   bool    isConnected();
+  uint8_t getAddress();
 
   //  RESET
   uint8_t reset();    //  reset both channels to AD524X_MIDPOINT and O1/O2 to LOW